% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{f.dist}
\alias{f.dist}
\title{Calculate within-node distance after splitting.}
\usage{
f.dist(threshold, raw.data, norm.data, sample.idx, test.var)
}
\arguments{
\item{threshold:}{threshold used to split the node}

\item{raw.data:}{Data that will be split.}

\item{norm.data:}{Data that will be split. Z-transformation has been applied.}

\item{sample.idx:}{row indices of samples in raw.data and norm.data, for which the splitting is optimized.}

\item{test.var:}{name of the variable, for which threshold will be shifted.}
}
\value{
mean of within-node distance.
}
\description{
Calculate within-node distance after splitting.
}
