% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.branch}
\alias{my.branch}
\title{split the data once. Among all N features, randomly sample sqrt(N) features, and choose one as the spliting feature. Selected is based on decision tree.}
\usage{
my.branch(data.train.sub, minsplit = 2, minbucket = 1, verbose = F)
}
\arguments{
\item{data.train.sub:}{A data frame with column "y" containing class labels and the remaining columns containing features.}

\item{minsplit:}{mininum number of genes in a node in order for a split to be attempted. Default value is 2.}

\item{minbucket:}{mininum number of genes in a node after splitting. Default value is 1.}
}
\value{
A decision tree with only one split.
}
\description{
split the data once. Among all N features, randomly sample sqrt(N) features, and choose one as the spliting feature. Selected is based on decision tree.
}
