% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.perf}
\alias{my.perf}
\title{Calculate performance using out-of-bag samples.}
\usage{
my.perf(forest)
}
\arguments{
\item{forest:}{a random forest model.}
}
\value{
A list containing two elements. The first element is a data frame with rows correponding to out-of-bag sample and columns correponding to correction predictions. The second element is the overall out-of-bag accuracy.
}
\description{
Calculate performance using out-of-bag samples.
}
