% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.predict}
\alias{my.predict}
\title{Make predictions using a random forest.}
\usage{
my.predict(forest, newdata)
}
\arguments{
\item{forest:}{a random forest model.}

\item{newdata:}{A data frame with genes as rows and features as columns. Prediction is made for each observation.}
}
\value{
A data frame with genes in rows and prediction values in columns. Columns contain prediction probabilities for a gene as a PG, OG, or TSG, respectively, predicted class label with the highest probability.
}
\description{
Make predictions using a random forest.
}
