% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.prune}
\alias{my.prune}
\title{Transfer learning by progressive pruning trees in a random forest.}
\usage{
my.prune(tree, newdata)
}
\arguments{
\item{forest:}{a random forest model trained in the source domain.}

\item{newdata:}{A data frame with genes as rows and features as columns. This is the target domain.}
}
\value{
A list of decision tree.
}
\description{
Transfer learning by progressive pruning trees in a random forest.
}
