% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.rf}
\alias{my.rf}
\title{Implementation of random forest algorithm.}
\usage{
my.rf(
  predictors,
  classes,
  random.pct = 0.8,
  ntrees = 200,
  minsplit = 20,
  minbucket = round(minsplit/3),
  set.seed = F
)
}
\arguments{
\item{predictors:}{A data frame with genes in rows and features in columns.}

\item{classes:}{A vector of class labels.}

\item{random.pct:}{Percentage to random sample genes.}

\item{ntrees:}{Number of trees in the random forest. Default value is 200.}

\item{minsplit:}{mininum number of genes in a node in order for a split to be attempted. Default value is 20.}

\item{minbucket:}{mininum number of genes in a node after splitting. Default value is round(minsplit/3).}
}
\value{
A list of decision trees.
}
\description{
Implementation of random forest algorithm.
}
