% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.rf.transfer}
\alias{my.rf.transfer}
\title{Transfer learning to adapt a random forest.}
\usage{
my.rf.transfer(forest, newdata, olddata, method = "both")
}
\arguments{
\item{forest:}{a random forest model trained in the source domain.}

\item{olddata:}{A data frame with genes as rows and features as columns. This is the source domain.}

\item{newdata:}{A data frame with genes as rows and features as columns. This is the target domain.}

\item{method:}{Adaption via "prune" (progressive pruning), "threshold" (threshold shifting), or "both'.}
}
\value{
A list of decision tree.
}
\description{
Transfer learning to adapt a random forest.
}
