% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.threshold}
\alias{my.threshold}
\title{Transfer learning by threshold shifting trees in a random forest.}
\usage{
my.threshold(tree, olddata, newdata)
}
\arguments{
\item{forest:}{a random forest model trained in the source domain.}

\item{olddata:}{A data frame with genes as rows and features as columns. This is the source domain.}

\item{newdata:}{A data frame with genes as rows and features as columns. This is the target domain.}
}
\value{
A list of decision tree.
}
\description{
Transfer learning by threshold shifting trees in a random forest.
}
