% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.transfer-learning.R
\name{my.tree}
\alias{my.tree}
\title{Build a decision tree with random subsamples of genes and features selected from random subsamples of features.}
\usage{
my.tree(
  predictors,
  classes,
  random.pct = 0.8,
  minsplit = 20,
  minbucket = round(minsplit/3),
  seed = NULL,
  verbose = F
)
}
\arguments{
\item{predictors:}{A data frame with genes in rows and features in columns.}

\item{classes:}{A vector of class labels.}

\item{random.pct:}{Perentage to random sample genes.}

\item{minsplit:}{mininum number of genes in a node in order for a split to be attempted. Default value is 20.}

\item{minbucket:}{mininum number of genes in a node after splitting. Default value is round(minsplit/3).}
}
\value{
A decision trees.
}
\description{
Build a decision tree with random subsamples of genes and features selected from random subsamples of features.
}
