% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-functions.selection.R
\name{transfer.rf}
\alias{transfer.rf}
\title{Transfer learning to adapt the model trained on human data to mouse data.}
\usage{
transfer.rf(output.folder, output.prefix, alt.fraction = 0.2)
}
\arguments{
\item{output.folder:}{A folder containing a "reformatted.txt" file.}

\item{output.prefix:}{A prefix string used to label the files. Transfered model is written to a file "md_transferred.RData".}
}
\description{
Transfer learning to adapt the model trained on human data to mouse data.
}
\examples{
: transfer.rf(output.folder='./examples/', output.prefix='mouse');
}
